backend/aits_backend/requirements.txt
anyio==4.4.0
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cohere==2.9.1
colorama==0.4.6
contourpy==1.3.0
cryptography==42.0.8
cycler==0.12.1
dj-database-url==2.3.0
Django==5.0.6
django-allauth==0.63.3
django-cors-headers==4.7.0
django-crispy-forms==2.2
django-environ==0.12.0
django-filter==24.2
django-rest-framework==0.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.6.1
fastapi==0.88.0
fonttools==4.54.1
gunicorn==23.0.0
h11==0.14.0
idna==3.7
Jinja2==3.1.4
kiwisolver==1.4.7
loguru==0.7.2
MarkupSafe==2.1.5
matplotlib==3.9.2
mysqlclient==2.2.5
numpy==2.1.3
packaging==24.2
pillow==11.0.0
pinecone-client==2.0.13
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==1.10.16
PyJWT==2.8.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.0
starlette==0.22.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.20.0
whitenoise==6.8.2
win32-setctime==1.1.0
